backend/attendance-kiosk/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
